+++ models/reports.go
@@ -9,10 +9,10 @@ type StoredReport struct {
 	Beneficiary                string  `json:"beneficiary"`
 	StartDate                  string  `json:"start_date"`
 	EndDate                    string  `json:"end_date"`
-	ContractAmount             float32 `json:"contract_amount"`
-	DiscountAmount             float32 `json:"discount_amount"`
+	ContractAmount             float64 `json:"contract_amount"`
+	DiscountAmount             float64 `json:"discount_amount"`
 	Content                    []byte  `json:"content"`
-	ContractAmountWithDiscount float32 `json:"contract_amount_with_discount"`
+	ContractAmountWithDiscount float64 `json:"contract_amount_with_discount"`
 	CreatedAt                  string  `json:"created_at"`
 	Author                     string  `json:"author"`
 }
